MACOS: improve build scripts

+++ macos/clean-rebuild.sh
@@ -8,12 +8,16 @@ git pull
 popd
 
 pushd ../../cross/spectmorph/macos
-rm -rf prefix build
-./build-deps.sh x86_64 clean
+if [ "x$1" != "xnodeps" ]; then
+  rm -rf prefix build
+  ./build-deps.sh x86_64 clean
+fi
 ./build.sh x86_64 clean
 popd
 
-rm -rf prefix build
-./build-deps.sh aarch64 clean
+if [ "x$1" != "xnodeps" ]; then
+  rm -rf prefix build
+  ./build-deps.sh aarch64 clean
+fi
 ./build.sh aarch64 clean
 ./dist.sh
